AddonManager: Consolidate preference in user.cfg from Plugins to  Addons
Currently when a user first uses the Addon Manager a dialog popup
prompts them about how FC is not liable for the code plugin authors
provide. It stores the binary response in a `readWarning` boolean
within the user.cfg file.

This commit changes the location of said boolean from
`User parameter:Plugins` to `User parameter:Addons`.

It also migrates old the old setting if available.

Co-authored-by: Benjamin Nauck <benjamin@nauck.se>