Default PCL to OFF on MSVC as well
PCL was already defaulted to OFF on Linux, this changes the default to OFF on MSVC
as well. This can be reverted once the primary LibPack for Windows includes a version
of FLANN that compiles under C++17 (the last official release of FLANN, 1.9.1, does
not, but the HEAD in their Git repository does).