Core: Allow users to add groups to active objects
As the title says, if right now there is Arch type active (like Level,
Building, etc. etc.), then it's not possible to assign Group to it
automatically (it's being created on root level of the document).

So this patch basically takes an active object and tries to insert it in
the same priority that Draft's autogroup functionality has.